Added missing return statements and move code into child component
